src/core/video_acquisition.py
"""Video Acquisition Module

This module handles video acquisition from multiple sources including streaming
platforms, TV broadcasts, YouTube, and other sources.
"""

import logging
from enum import Enum
from typing import List, Dict, Any, Optional, Union

import ffmpeg
import requests
from pydantic import BaseModel

logger = logging.getLogger(__name__)


class ContentSource(str, Enum):
    """Enum representing different content sources."""
    STREAMING = "streaming"  # Netflix, Disney+, etc.
    TELEVISION = "television"  # Network television
    YOUTUBE = "youtube"  # YouTube content
    ARCHIVE = "archive"  # Archive content
    USER = "user"  # User-submitted content


class SourceTier(int, Enum):
    """Tier classification for content sources."""
    TIER_1 = 1  # Popular streaming platforms top 10,000 titles
    TIER_2 = 2  # Network television shows (current season)
    TIER_3 = 3  # YouTube trending content and channels >1M subscribers
    TIER_4 = 4  # Archive content based on popularity metrics


class VideoMetadata(BaseModel):
    """Metadata for acquired videos."""
    source_id: str
    title: str
    source: ContentSource
    tier: SourceTier
    duration: float
    resolution: str
    format: str
    acquisition_timestamp: str
    additional_metadata: Dict[str, Any] = {}


class VideoAcquisitionService:
    """Service for acquiring videos from various sources."""
    
    def __init__(self, config: Dict[str, Any]):
        """Initialize the video acquisition service.
        
        Args:
            config: Configuration dictionary for the service
        """
        self.config = config
        self.api_keys = config.get("api_keys", {})
        logger.info("Initialized VideoAcquisitionService")
    
    async def acquire_from_streaming(self, platform: str, content_id: str) -> Optional[Dict[str, Any]]:
        """Acquire video from streaming platforms.
        
        Args:
            platform: Streaming platform name (netflix, disney, etc.)
            content_id: ID of the content to acquire
            
        Returns:
            Dictionary with video data and metadata, or None if acquisition failed
        """
        logger.info(f"Acquiring video from {platform}, content ID: {content_id}")
        # Implementation would depend on platform-specific APIs or methods
        # This is a placeholder for the actual implementation
        return {"status": "success", "path": f"/storage/{platform}/{content_id}.mp4", 
                "metadata": {"title": f"Content {content_id}", "platform": platform}}
    
    async def acquire_from_television(self, channel: str, timestamp: str, duration: int) -> Optional[Dict[str, Any]]:
        """Acquire video from television broadcast.
        
        Args:
            channel: TV channel name or ID
            timestamp: Broadcast timestamp to acquire
            duration: Duration in seconds to acquire
            
        Returns:
            Dictionary with video data and metadata, or None if acquisition failed
        """
        logger.info(f"Acquiring {duration}s of TV content from {channel} at {timestamp}")
        # Implementation would connect to TV recording systems
        # This is a placeholder for the actual implementation
        return {"status": "success", "path": f"/storage/tv/{channel}/{timestamp}.mp4", 
                "metadata": {"channel": channel, "broadcast_time": timestamp}}

    async def acquire_from_youtube(self, video_id: str) -> Optional[Dict[str, Any]]:
        """Acquire video from YouTube.
        
        Args:
            video_id: YouTube video ID
            
        Returns:
            Dictionary with video data and metadata, or None if acquisition failed
        """
        youtube_api_key = self.api_keys.get("youtube")
        if not youtube_api_key:
            logger.error("YouTube API key not found in configuration")
            return None
            
        logger.info(f"Acquiring YouTube video: {video_id}")
        # This would use YouTube API or a library like pytube
        # This is a placeholder for the actual implementation
        return {"status": "success", "path": f"/storage/youtube/{video_id}.mp4", 
                "metadata": {"video_id": video_id, "platform": "youtube"}}

    async def process_ingestion_queue(self, queue_items: List[Dict[str, Any]]) -> List[Dict[str, Any]]:
        """Process a batch of items from the ingestion queue.
        
        Args:
            queue_items: List of items to process from the queue
            
        Returns:
            List of results for each processed item
        """
        results = []
        for item in queue_items:
            source = item.get("source")
            content_id = item.get("content_id")
            
            if source == ContentSource.STREAMING:
                result = await self.acquire_from_streaming(item.get("platform"), content_id)
            elif source == ContentSource.TELEVISION:
                result = await self.acquire_from_television(
                    item.get("channel"), item.get("timestamp"), item.get("duration", 3600)
                )
            elif source == ContentSource.YOUTUBE:
                result = await self.acquire_from_youtube(content_id)
            else:
                logger.warning(f"Unsupported content source: {source}")
                result = {"status": "error", "error": "Unsupported content source"}
                
            results.append({"item": item, "result": result})
            
        return results
