requirements.txt
# Core dependencies
python-dotenv>=0.19.2
pydantic>=1.9.0
fastapi>=0.78.0
uvicorn>=0.17.6
sqlalchemy>=1.4.36
psycopg2-binary>=2.9.3
alembic>=1.8.0

# Video processing
opencv-python>=4.5.5
ffmpeg-python>=0.2.0
scenedetect>=0.6.0
imageio>=2.19.3
imageio-ffmpeg>=0.4.7

# Feature extraction
torch>=1.11.0
torchvision>=0.12.0
librosa>=0.9.1
h5py>=3.7.0
numpy>=1.22.3
scipy>=1.8.0
sklearn>=0.0
scikit-learn>=1.1.1
scikit-image>=0.19.2

# Vector database clients
pymilvus>=2.1.0
pinecone-client>=2.0.0

# Distributed processing
apache-beam>=2.38.0
ray>=1.13.0

# API and Web
starlette>=0.19.1
httpx>=0.22.0
jinja2>=3.1.2
aiofiles>=0.8.0
python-multipart>=0.0.5
itsdangerous>=2.1.2
pyjwt>=2.4.0

# Testing
pytest>=7.1.2
pytest-cov>=3.0.0

# Development tools
black>=22.3.0
isort>=5.10.1
flake8>=4.0.1
mypy>=0.950
pre-commit>=2.19.0

# Cloud clients
awscli>=1.24.0
boto3>=1.23.10
google-cloud-storage>=2.3.0
azure-storage-blob>=12.12.0

# Monitoring and logging
prometheus-client>=0.14.1
loguru>=0.6.0
